agent/requirements.txt
# Python dependencies for Budget Agent

# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35

# Google Generative AI
google-generativeai==0.8.3

# Data processing
pandas==2.2.3

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2

